marketplace_service/requirements.txt
Django==4.2.2
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
gunicorn==21.2.0
djangorestframework==3.14.0
daphne==4.0.0
django_filter==23.2
django-phonenumber-field==8.0.0
Pillow==11.0.0